rename: VO -> enums
