+++ src/main/java/softeer2nd/domain/enums/Color.java
@@ -1,4 +1,4 @@
-package softeer2nd.domain.VO;
+package softeer2nd.domain.enums;
 
 public enum Color {
     WHITE,
