+++ src/test/java/softeer2nd/domain/PieceTest.java
@@ -2,8 +2,8 @@ package softeer2nd.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.domain.VO.Color;
-import softeer2nd.domain.VO.Type;
+import softeer2nd.domain.enums.Color;
+import softeer2nd.domain.enums.Type;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
