+++ src/test/java/softeer2nd/domain/RankTest.java
@@ -3,8 +3,8 @@ package softeer2nd.domain;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.domain.VO.Color;
-import softeer2nd.domain.VO.Type;
+import softeer2nd.domain.enums.Color;
+import softeer2nd.domain.enums.Type;
 
 import static org.junit.jupiter.api.Assertions.*;
